% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{fitted.multippmnet}
\alias{fitted.multippmnet}
\title{Fitted intensity from a 'multippmnet' object}
\usage{
\method{fitted}{multippmnet}(object, s = NULL, drop = FALSE, ...)
}
\description{
Computes the fitted intensity for a regularized inhomogeneous Poisson point
process model at the quadrature points used to fit the model.
}
