% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multippmnet.R
\name{multippmnet}
\alias{multippmnet}
\title{Fit a regularized inhomogeneous multivariate Poisson point process model}
\usage{
multippmnet(Q, data, ..., grp_weights = NULL, ind_weights = NULL)
}
\arguments{
\item{Q}{A quadrature scheme (of class \code{quad}) containing a multivariate
point pattern.}

\item{data}{A list of pixel images (of class \code{imlist}) representing
spatial covariates.}

\item{...}{Additional arguments passed to \code{asgl} to control the fitting
procedure.}

\item{grp_weights}{Optional. A vector of adaptive weights applied to the
groups of coefficients associated with each covariate in \code{data}.}

\item{ind_weights}{Optional. A vector of adaptive weights applied to the
individual model coefficients.}
}
\value{
An object of class \code{multippmnet}.
}
\description{
Fit an inhomogeneous multivariate Poisson point process model via penalized
conditional logistic regression. An adaptive sparse group lasso penalty is
employed in order to exploit grouped structure in the model coefficients.
}
